+++ utils/user.go
@@ -11,7 +11,7 @@ import (
 )
 
 //GetUserID gets the user id from a web request
-func GetUserID(r *http.Request) (int, error) {
+func GetUserID(r *http.Request) (int64, error) {
 	session := r.Header.Get("Authorization")
 	if len(session) == 0 {
 		//Check if maybe cookie was set
